extract/requirements.txt
requests>=2.31.0
boto3>=1.40.0
python-dotenv>=1.1.0